requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
fastapi==0.120.4
python-multipart==0.0.20
uvicorn[standard]==0.38.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.0
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.9
MarkupSafe==3.0.3
marl==1.0.0
matplotlib==3.10.7
monai==1.5.1
mpmath==1.3.0
networkx==3.2.1
numpy==2.3.4
openai==1.91.0
packaging==25.0
pandas==2.3.3
pettingzoo==1.24.3
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.0
pydantic==2.12.0
pydantic_core==2.41.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tifffile==2025.10.16
torch==2.9.0
torchvision==0.24.0
torchxrayvision==1.4.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
